fix: only show instance license stats for admin users
